[Qt] Add quirks for running the web process in a profiler shell, like valgrind https://bugs.webkit.org/show_bug.cgi?id=87672
Reviewed by Jocelyn Turcotte.

Source/WebKit2:

If environment variables QT_WEBKIT_WEB_PROCESS_COMMAND_PREFIX
and/or QT_WEBKIT_PLUGIN_PROCESS_COMMAND_PREFIX are set, use their
values as the prefix of the command when launching the child process.
Example usage:
    export QT_WEB_PROCESS_COMMAND_PREFIX="valgrind --smc-check=all"
    ./MiniBrowser http://somepage.html
    ./WebKitTestRunner --no-timeout-at-all some_layouttest.html

* UIProcess/Launcher/qt/ProcessLauncherQt.cpp:
(WebKit::ProcessLauncher::launchProcess):

Tools:

Add --no-timeout-at-all command line switch to not time out no matter how
long the web process does it's job. --no-timeout is not enough for profiling
purposes because it only affects the timeout when we are waiting for the end
of the test. We use another timeout to control how much we wait for the web
process to initialize before and reset after the test and it could also easily
expire when running under valgrind. We should keep the behavior of --no-timeout
to not disable this timout check because --no-timeout is used by performace tests
and those should not wait for the web process to initialize and reset more than
normal tests.

* WebKitTestRunner/TestController.cpp:
(WTR::TestController::TestController):
(WTR::TestController::initialize):
(WTR::TestController::runUntil):
* WebKitTestRunner/TestController.h:

Identifier: 109914@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123464 268f45cc-cd09-0410-ab3c-d52691b4dbfc