Crash in RenderTableSection::layoutRows https://bugs.webkit.org/show_bug.cgi?id=92145
Reviewed by Abhishek Arya.

Source/WebCore:

The crash came from the insertChildNode logic being confused and choosing the wrong
renderer for insertion. This is because after r123159, table anonymous wrappers can
be created as 'inline-table' and wouldn't pass the isAnonymousBlock() check.

This change also aligns the ancestor iteration with RenderBlock::addChildToAnonymousColumnBlocks.

Test: fast/table/crash-table-section-layoutRows.html

* rendering/RenderObjectChildList.cpp:
(WebCore::RenderObjectChildList::insertChildNode):
Changed the code to always get the right child. There is no justification to not
walking up to the |owner|'s direct child in all cases.

LayoutTests:

* fast/table/crash-table-section-layoutRows-expected.txt: Added.
* fast/table/crash-table-section-layoutRows.html: Added.


Identifier: 109972@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123525 268f45cc-cd09-0410-ab3c-d52691b4dbfc