[Qt][WK2] Disable CopyToTexture for GraphicsSurfaceMac. https://bugs.webkit.org/show_bug.cgi?id=92159
Disable CopyToTexture feature for GraphicsSurface on Mac.
While this is actually working, it is currently used for Tiles
that are single buffered, and therefore requires a call to glFlush.
This call blocks the GPU for about 40ms which would make smooth animations impossible.

Reviewed by Noam Rosenthal.

* platform/graphics/surfaces/mac/GraphicsSurfaceMac.cpp:
(WebCore::GraphicsSurface::platformCreate):
(WebCore::GraphicsSurface::platformImport):

Identifier: 110028@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123588 268f45cc-cd09-0410-ab3c-d52691b4dbfc