Search cancel button is hard to activate with a tap gesture even if touch adjustment is enabled. https://bugs.webkit.org/show_bug.cgi?id=91894
Patch by Kevin Ellis <kevers@chromium.org> on 2012-07-27
Reviewed by Antonio Gomes.

Source/WebCore:

Update check for determining if a node responds to tap gestures.
Previously, the cancel button was being discarded as a candidate for
touch adjustment if the touch area clipped the text input region
because the candidate pruning process failed to account for the cancel
button having a default mouse handler.  Possibly an interim fix,
depending on resolution to 92093.

Test: touchadjustment/search-cancel.html

* page/TouchAdjustment.cpp:
(TouchAdjustment):
(WebCore::TouchAdjustment::nodeRespondsToTapGesture):

LayoutTests:

Add test for touch adjustment in the vicinity of a search cancel button.
Prior to the patch, the cancel button was pruned as a candidate if the
touch area clipped the input field.

* touchadjustment/search-cancel-expected.txt: Added.
* touchadjustment/search-cancel.html: Added.

Identifier: 110354@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123919 268f45cc-cd09-0410-ab3c-d52691b4dbfc