Wheel events on a page with frames are not handled in fixed layout https://bugs.webkit.org/show_bug.cgi?id=90550
Reviewed by Antonio Gomes.

Based on patch by Andras Becsi.

Frame-rect size is not the right value to check against to see if the document
can be contained within the full visible content rect, since it does not handle
fixedVisibleContentRect. Instead visibleContentRect(true) should be used which
returns the full visible content rect disregarding scrollbars, and which is
the same a frame-rect size when fixedVisibleContentRect is not set.

* platform/ScrollView.cpp:
(WebCore::ScrollView::updateScrollbars):

Identifier: 110458@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124024 268f45cc-cd09-0410-ab3c-d52691b4dbfc