ASSERTION FAILED: !rect.isEmpty()  : void WebCore::GraphicsContext::drawRect(const WebCore::IntRect &) https://bugs.webkit.org/show_bug.cgi?id=92187
Reviewed by Simon Fraser.

Source/WebCore:

The border painting logic (RenderBoxModelObject::paintOneBorderSide) would pass a rect with a 0px length
to RenderObject::drawLineForBoxSide. We do check the width (thickness) but not the length so we would pass
the rect to GraphicsContext and hit the ASSERT. This change adds a check for the length too as it is the safest
way, it means that we may still do unneeded operations before bailing out but that's an existing problem in the code.

Tests: fast/borders/0px-borders-no-line-height.html
       fast/borders/double-1px-border-assert.html

* rendering/RenderObject.cpp:
(WebCore::RenderObject::drawLineForBoxSide):
Added a 0px length check. While renaming confusing variables and re-using others,
I also found a potential empty border that I fixed (tested by the 2nd case above).

LayoutTests:

* fast/borders/0px-borders-no-line-height-expected.html: Added.
* fast/borders/0px-borders-no-line-height.html: Added.
* fast/borders/double-1px-border-assert-expected.html: Added.
* fast/borders/double-1px-border-assert.html: Added.


Identifier: 110473@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124044 268f45cc-cd09-0410-ab3c-d52691b4dbfc